# Get Parquet string reading offsets in C++ function