Previously, Parquet string files that had been read in were
calculating their offset arrays with the `segmentedCalcOffsets`
function from `GenSymIO.chpl`, but that function is not optimized
and was resulting in large amounts of communication, meaning that
performance decreased as more nodes were added.

This PR switches the computation of the offsets array to happen
at the point where the byte sizes are determined from the file,
resulting in much better scaling.

Here are some performance numbers calculated on a CS using
50 files of 1 million uniformly distributed string elements:

nodes | master  | this branch |
----- | ------: | ----------: |
1     | 2.366 s | 0.3248 s    |
2     | minutes | 0.2899 s    |
4     | minutes | 0.2625 s    |
8     | minutes | 0.2135 s    |